requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
docutils==0.21.2
filetype==1.2.0
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.10
pwinput==1.0.3
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
urllib3==2.3.0
